requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
executing==1.2.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kombu==5.2.4
magic-filter==1.0.9
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
pyzmq==25.0.0
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
